+++ src/app/navbar.tsx
@@ -1,0 +1,41 @@
+'use client'
+
+import { Navbar, NavbarBrand, NavbarContent, NavbarItem, Link, Button } from '@nextui-org/react'
+import React, { ReactNode } from 'react'
+
+export default function NavigateBar({ children }: { children: ReactNode }) {
+  return (
+    <>
+      <Navbar>
+        <NavbarBrand>
+          <p className='font-bold text-inherit'>
+            <a href='/'>Admin Console Boilerplate</a>
+          </p>
+        </NavbarBrand>
+        <NavbarContent className='hidden sm:flex gap-4' justify='center'>
+          <NavbarItem>
+            <Link color='foreground' href='/nested'>
+              Nested
+            </Link>
+          </NavbarItem>
+          <NavbarItem isActive>
+            <Link href='#' aria-current='page'>
+              Settings
+            </Link>
+          </NavbarItem>
+        </NavbarContent>
+        <NavbarContent justify='end'>
+          <NavbarItem className='hidden lg:flex'>
+            <Link href='#'>Login</Link>
+          </NavbarItem>
+          <NavbarItem>
+            <Button as={Link} color='primary' href='#' variant='flat'>
+              Sign Up
+            </Button>
+          </NavbarItem>
+        </NavbarContent>
+      </Navbar>
+      {children}
+    </>
+  )
+}
